refactor(bridge): move callback metadata synchronization
